Implemented QueryType override
Any QueryType class can be registered as a new instance
by specifying the 'alias' tag attribute:

```
my_override_querytype:
  class: AppBundle\QueryType\MyQueryType
  arguments:
    - 'some_argument'
  tags:
    - {name: ezpublish.query_type, alias: my_override}

my_second_override_querytype:
  class: AppBundle\QueryType\MyQueryType
  arguments:
    - 'some_other_argument'
  tags:
    - {name: ezpublish.query_type, alias: my_second_override}
```

It makes it easy to add variations of an existing querytype with
a different set of arguments or services. If it is named by convention,
the overridden querytype is still registered under its own name,
based on the static method.